api integration and UI except order management

+++ app/Http/Controllers/ItemController.php
@@ -3,11 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\item;
+use App\Libraries\AclHandler;
 use App\Libraries\HandleApi;
 use App\shop;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Yajra\DataTables\DataTables;
+use Intervention\Image\Facades\Image as Image;
 
 class ItemController extends Controller
 {
@@ -18,6 +20,10 @@ class ItemController extends Controller
      */
     public function index()
     {
+        if (AclHandler::hasAccess('Product','full') == false){
+            die('Not access . Recorded this '); exit();
+        }
+
         return view('item.add_item');
     }
 
@@ -28,6 +34,10 @@ class ItemController extends Controller
      */
     public function itemList()
     {
+        if (AclHandler::hasAccess('Product','full') == false){
+            die('Not access . Recorded this '); exit();
+        }
+
         $brand_api_url = "https://xplaza-backend.herokuapp.com/api/brand";
         $brandCurlOutput  = HandleApi::getCURLOutput( $brand_api_url, 'GET', [] );
         $brand_json_resp = json_decode($brandCurlOutput);
@@ -43,7 +53,17 @@ class ItemController extends Controller
         $shop_json_resp = json_decode($shopCurlOutput);
         $shops = isset($shop_json_resp->data) ? $shop_json_resp->data : [];
 
-        return view('item.item_list',compact('brands','categories','shops'));
+        $shop_api_url = "https://xplaza-backend.herokuapp.com/api/currency";
+        $shopCurlOutput  = HandleApi::getCURLOutput( $shop_api_url, 'GET', [] );
+        $shop_json_resp = json_decode($shopCurlOutput);
+        $currencies = isset($shop_json_resp->data) ? $shop_json_resp->data : [];
+
+        $shop_api_url = "https://xplaza-backend.herokuapp.com/api/prodvartype";
+        $shopCurlOutput  = HandleApi::getCURLOutput( $shop_api_url, 'GET', [] );
+        $shop_json_resp = json_decode($shopCurlOutput);
+        $prodvartypes = isset($shop_json_resp->data) ? $shop_json_resp->data : [];
+
+        return view('item.item_list',compact('brands','categories','shops','currencies','prodvartypes'));
     }
 
 
@@ -57,9 +77,16 @@ class ItemController extends Controller
 
         return Datatables::of(collect($data))
             ->addColumn('action', function ($data) {
-                $action = '<button type="button" class="btn btn-info btn-xs open_category_modal" data-item_id="'.$data->id.'" ><b><i class="fa fa-edit"></i> Edit</b></button> &nbsp;';
-                $action .= ' <button type="button" class="btn btn-danger btn-xs deleteItem" data-item_id="'.$data->id.'"><b><i class="fa fa-trash"></i> Delete</b></button>';
-
+                $action = '';
+                if (AclHandler::hasAccess('Product','view') == true) {
+                    $action = '<button type="button" class="btn btn-primary btn-xs open_item_view_modal" data-item_id="' . $data->id . '" ><b><i class="fa fa-eye"></i> View</b></button> &nbsp;';
+                }
+                if (AclHandler::hasAccess('Product','update') == true) {
+                    $action .= '<button type="button" class="btn btn-info btn-xs open_item_modal" data-item_id="' . $data->id . '" ><b><i class="fa fa-edit"></i> Edit</b></button> &nbsp;';
+                }
+                if (AclHandler::hasAccess('Product','delete') == true) {
+                    $action .= ' <button type="button" class="btn btn-danger btn-xs deleteItem" data-item_id="' . $data->id . '"><b><i class="fa fa-trash"></i> Delete</b></button>';
+                }
                 return $action;
             })
             ->removeColumn('id')
@@ -84,7 +111,223 @@ class ItemController extends Controller
      */
     public function store(Request $request)
     {
+        if (AclHandler::hasAccess('Product','add') == false){
+            return response()->json( ['responseCode'=>0,'message'=>'Not access . Recorded this']);
+        }
+
         $rules = [
+            'item_name'    => 'required',
+            'item_image'    => 'required',
+            'quantity'    => 'required',
+            'description'    => 'required',
+            'shop_id'    => 'required',
+            'brand_id'    => 'required',
+            'category_id'    => 'required',
+            'buying_price'    => 'required',
+            'selling_price'    => 'required',
+            'currency_id'    => 'required',
+            'product_var_type_id'    => 'required',
+            'product_var_type_option'    => 'required'
+        ];
+        $validator = Validator::make( $request->all(), $rules );
+        if ( $validator->fails() ) {
+           // return response()->json( ['responseCode'=>0,'message'=>'Please fill up required field']);
+        }
+
+        $item_name = $request->get('item_name');
+        $item_image = $request->get('item_image');
+        $description = $request->get('description');
+        $shop_id = $request->get('shop_id');
+        $quantity = $request->get('quantity');
+        $brand_id = $request->get('brand_id');
+        $category_id = $request->get('category_id');
+        $buying_price = $request->get('buying_price');
+        $selling_price = $request->get('selling_price');
+        $currency_id = $request->get('currency_id');
+        $product_var_type_id = $request->get('product_var_type_id');
+        $product_var_type_option = $request->get('product_var_type_option');
+
+        list($type, $data) = explode(';', $item_image);
+        list(, $data) = explode(',', $data);
+        $image     = Image::make(base64_decode($data))->encode('jpg');
+        $imageName = date('ymdhis') . '.jpg';
+        $image->save(public_path() . '/item_image/' . $imageName);
+
+        $bodyData = [
+            "brand_id"=>$brand_id,
+            "buying_price"=>$buying_price ,
+            "category_id"=>$category_id,
+            "quantity"=>$quantity,
+            "currency_id"=>$currency_id,
+            "description"=>$description,
+            "name"=>$item_name,
+            "product_var_type_id"=>$product_var_type_id,
+            "product_var_type_option"=>$product_var_type_option,
+            "selling_price"=>$selling_price,
+            "shop_id"=>$shop_id
+        ];
+        $fieldData = json_encode($bodyData);
+
+        $api_url = "https://xplaza-backend.herokuapp.com/api/product/add";
+        $curlOutputMain  = HandleApi::getCURLOutput( $api_url, 'POST', $fieldData );
+
+        $decodedResp = json_decode($curlOutputMain);
+        if($decodedResp->status == 201){
+            if(isset($decodedResp->data)){
+                preg_match_all('!\d+!', $decodedResp->data, $numberOnly);
+                if(isset($numberOnly[0][0])){
+                    $bodyData = [
+                        "name"=>$imageName,
+                        "path"=>'',
+                        "product_id"=>$numberOnly[0][0]
+                    ];
+                    $fieldData = json_encode($bodyData);
+                    $api_url = "https://xplaza-backend.herokuapp.com/api/productimage/add";
+                    $curlOutput  = HandleApi::getCURLOutput( $api_url, 'POST', $fieldData );
+                 //   dd($curlOutput);
+                }
+            }
+        }
+
+
+        $decodedResp = json_decode($curlOutputMain);
+        if($decodedResp->status == 201){
+            return response()->json( ['responseCode'=>1,'message'=>'Successfully added']);
+        }else{
+            return response()->json( ['responseCode'=>0,'message'=>$decodedResp->message]);
+        }
+
+
+    }
+
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+
+    public function itemInfo(Request $request)
+    {
+        $rules = [
+            'item_id'        => 'required'
+        ];
+        $validator = Validator::make( $request->all(), $rules );
+        if ( $validator->fails() ) {
+            return response()->json( ['responseCode'=>0,'message'=>'Please fill up required field']);
+        }
+
+        $item_id = $request->get('item_id');
+
+        $brand_api_url = "https://xplaza-backend.herokuapp.com/api/brand";
+        $brandCurlOutput  = HandleApi::getCURLOutput( $brand_api_url, 'GET', [] );
+        $brand_json_resp = json_decode($brandCurlOutput);
+        $brands = isset($brand_json_resp->data) ? $brand_json_resp->data : [];
+
+        $category_api_url = "https://xplaza-backend.herokuapp.com/api/category";
+        $categoryCurlOutput  = HandleApi::getCURLOutput( $category_api_url, 'GET', [] );
+        $category_json_resp = json_decode($categoryCurlOutput);
+        $categories = isset($category_json_resp->data) ? $category_json_resp->data : [];
+
+        $shop_api_url = "https://xplaza-backend.herokuapp.com/api/shop";
+        $shopCurlOutput  = HandleApi::getCURLOutput( $shop_api_url, 'GET', [] );
+        $shop_json_resp = json_decode($shopCurlOutput);
+        $shops = isset($shop_json_resp->data) ? $shop_json_resp->data : [];
+
+        $api_url = "https://xplaza-backend.herokuapp.com/api/product/".intval($item_id);
+        $curlOutput  = HandleApi::getCURLOutput( $api_url, 'GET', [] );
+        $decodedData = json_decode($curlOutput);
+        $item_data = isset($decodedData->data) ? $decodedData->data : [];
+
+        $api_url = "https://xplaza-backend.herokuapp.com/api/product/".intval($item_id);
+        $curlOutputMain  = HandleApi::getCURLOutput( $api_url, 'GET', [] );
+        $decodedDataForItem = json_decode($curlOutputMain);
+        $itemInfo = $decodedDataForItem->data;
+
+        $shop_api_url = "https://xplaza-backend.herokuapp.com/api/currency";
+        $shopCurlOutput  = HandleApi::getCURLOutput( $shop_api_url, 'GET', [] );
+        $shop_json_resp = json_decode($shopCurlOutput);
+        $currencies = isset($shop_json_resp->data) ? $shop_json_resp->data : [];
+
+        $shop_api_url = "https://xplaza-backend.herokuapp.com/api/prodvartype";
+        $shopCurlOutput  = HandleApi::getCURLOutput( $shop_api_url, 'GET', [] );
+        $shop_json_resp = json_decode($shopCurlOutput);
+        $prodvartypes = isset($shop_json_resp->data) ? $shop_json_resp->data : [];
+
+        $api_url = "https://xplaza-backend.herokuapp.com/api/productimage/".intval($itemInfo->id);
+        $curlOutputMain  = HandleApi::getCURLOutput( $api_url, 'GET', [] );
+        $decodedDataForItem = json_decode($curlOutputMain);
+        $imagenfo = $decodedDataForItem->data;
+     //   dd($imagenfo,$itemInfo->id);
+
+        if(isset($imagenfo[0]->name)){
+            $image_url = $imagenfo[0]->name;
+            $image_id = $imagenfo[0]->id;
+
+        }else{
+            $image_id = '';
+            $image_url = '';
+        }
+
+        $public_html = strval(view("item.modal_data", compact('item_data','brands', 'categories', 'shops', 'itemInfo','image_url','currencies','prodvartypes','image_id')));
+
+        return response()->json(['responseCode' => 1, 'html' => $public_html,'message'=>'Successfully fetches']);
+
+    }
+
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function itemInfoView(Request $request)
+    {
+        $rules = [
+            'item_id'        => 'required'
+        ];
+        $validator = Validator::make( $request->all(), $rules );
+        if ( $validator->fails() ) {
+            return response()->json( ['responseCode'=>0,'message'=>'Please fill up required field']);
+        }
+
+        $item_id = $request->get('item_id');
+
+        $api_url = "https://xplaza-backend.herokuapp.com/api/product/".intval($item_id);
+        $curlOutputMain  = HandleApi::getCURLOutput( $api_url, 'GET', [] );
+        $decodedDataForItem = json_decode($curlOutputMain);
+        $itemInfo = $decodedDataForItem->data;
+
+        $api_url = "https://xplaza-backend.herokuapp.com/api/productimage/".intval($itemInfo->id);
+        $curlOutputMain  = HandleApi::getCURLOutput( $api_url, 'GET', [] );
+        $decodedDataForItem = json_decode($curlOutputMain);
+        $imagenfo = $decodedDataForItem->data;
+
+        if(isset($imagenfo[0]->name)){
+            $image_url = $imagenfo[0]->name;
+            $image_id = $imagenfo[0]->id;
+
+        }else{
+            $image_id = '';
+            $image_url = '';
+        }
+
+        $public_html = strval(view("item.modal_data_view", compact('itemInfo','image_url')));
+
+        return response()->json(['responseCode' => 1, 'html' => $public_html,'message'=>'Successfully fetches']);
+
+    }
+
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+
+    public function updateItem(Request $request)
+    {
+        if (AclHandler::hasAccess('Product','update') == false){
+            return response()->json( ['responseCode'=>0,'message'=>'Not access . Recorded this']);
+        }
+
+        $rules = [
+            'item_id'    => 'required',
+            'quantity'    => 'required',
             'item_name'    => 'required',
             'description'    => 'required',
             'shop_id'    => 'required',
@@ -98,10 +341,13 @@ class ItemController extends Controller
         ];
         $validator = Validator::make( $request->all(), $rules );
         if ( $validator->fails() ) {
-            return response()->json( ['responseCode'=>0,'message'=>'Please fill up required field']);
+            // return response()->json( ['responseCode'=>0,'message'=>'Please fill up required field']);
         }
 
+        $item_id = $request->get('item_id');
+        $quantity = $request->get('quantity');
         $item_name = $request->get('item_name');
+        $item_image = $request->get('item_image');
         $description = $request->get('description');
         $shop_id = $request->get('shop_id');
         $brand_id = $request->get('brand_id');
@@ -112,8 +358,11 @@ class ItemController extends Controller
         $product_var_type_id = $request->get('product_var_type_id');
         $product_var_type_option = $request->get('product_var_type_option');
 
+
         $bodyData = [
+            "id"=>$item_id,
             "brand_id"=>$brand_id,
+            "quantity"=>$quantity,
             "buying_price"=>$buying_price ,
             "category_id"=>$category_id,
             "currency_id"=>$currency_id,
@@ -126,44 +375,40 @@ class ItemController extends Controller
         ];
         $fieldData = json_encode($bodyData);
 
-        $api_url = "https://xplaza-backend.herokuapp.com/api/product/add";
-        $curlOutput  = HandleApi::getCURLOutput( $api_url, 'POST', $fieldData );
+        $api_url = "https://xplaza-backend.herokuapp.com/api/product/update";
+        $curlOutputMain  = HandleApi::getCURLOutput( $api_url, 'PUT', $fieldData );
 
-        return response()->json( ['responseCode'=>1,'message'=>'Successfully added']);
+        $decodedResp = json_decode($curlOutputMain);
+
+        if($decodedResp->status == 200 && isset($item_image)){
+            list($type, $data) = explode(';', $item_image);
+            list(, $data) = explode(',', $data);
+            $image     = Image::make(base64_decode($data))->encode('jpg');
+            $imageName = date('ymdhis') . '.jpg';
+            $image->save(public_path() . '/item_image/' . $imageName);
 
 
-    }
-
-    /**
-     * @param Request $request
-     * @return \Illuminate\Http\JsonResponse
-     */
-
-    public function categoryInfo(Request $request)
-    {
-        $rules = [
-            'category_id'        => 'required'
-        ];
-        $validator = Validator::make( $request->all(), $rules );
-        if ( $validator->fails() ) {
-            return response()->json( ['responseCode'=>0,'message'=>'Please fill up required field']);
+            $bodyData = [
+                "id" => intval($request->get('item_image_id')),
+                "name"=>$imageName,
+                "path"=>'',
+                "product_id"=>$item_id
+            ];
+            $fieldData = json_encode($bodyData);
+            $api_url = "https://xplaza-backend.herokuapp.com/api/productimage/update";
+            $curlOutput  = HandleApi::getCURLOutput( $api_url, 'PUT', $fieldData );
         }
 
-        $category_id = $request->get('category_id');
 
-        $bodyData = [
-            "category_id"=>$category_id
-        ];
-        $fieldData = json_encode($bodyData);
-
-        $api_url = "https://xplaza-backend.herokuapp.com/api/shop/add";
-        //  $curlOutput  = HandleApi::getCURLOutput( $api_url, 'POST', $fieldData );
-
-        return response()->json( ['responseCode'=>1,'message'=>'Successfully fetches']);
+        $decodedResp = json_decode($curlOutputMain);
+        if($decodedResp->status == 200){
+            return response()->json( ['responseCode'=>1,'message'=>'Successfully updated']);
+        }else{
+            return response()->json( ['responseCode'=>0,'message'=>$decodedResp->message]);
+        }
 
 
     }
-
 
     /**
      * @param Request $request
@@ -171,6 +416,11 @@ class ItemController extends Controller
      */
     public function deleteItem(Request $request)
     {
+
+        if (AclHandler::hasAccess('Product','delete') == false){
+            return response()->json( ['responseCode'=>0,'message'=>'Not access . Recorded this']);
+        }
+
         $rules = [
             'item_id'        => 'required'
         ];
@@ -184,7 +434,11 @@ class ItemController extends Controller
 
         $decodedData = json_decode($curlOutput);
 
-        return response()->json( ['responseCode'=>1,'message'=>'Successfully Deleted']);
+        if($decodedData->status == 200){
+            return response()->json( ['responseCode'=>1,'message'=>'Successfully updated']);
+        }else{
+            return response()->json( ['responseCode'=>0,'message'=>$decodedData->message]);
+        }
 
 
     }
